feat(admin): unify user management and clarify search actions
Merge the former Permissions screen into Manage Users (search, edit profile, groups, grant/revoke admin, plus administrators list). Remove the duplicate Permissions nav entry, redirect /admin/permissions to /admin/users, and replace row-click and icon actions with explicit Edit User and Edit Groups buttons alongside admin controls.

Made-with: Cursor
